fix some parts according to code review
